Added necessary path to example code
Hi, thanks for pulling this together! I just discovered an error that I'd seen in my own library dev. When code is including a library rather than having the files wrapped up with it in the editor then the library's files are actually in a subdirectory named identically to the library. Check my fix if that's a bit murky.

Example code compiles now!